fix NPE on PCGen load, update Agent instructions, stabalize some UI elements (row open/close)
