+++ code/src/java/pcgen/gui3/component/PCGenToolBar.java
@@ -26,7 +26,7 @@ import pcgen.gui2.PCGenFrame;
 import pcgen.gui2.PCGenUIManager;
 import pcgen.gui2.dialog.PrintPreviewDialog;
 import pcgen.gui2.tools.Icons;
-import pcgen.gui3.JFXPanelFromResource;
+import pcgen.gui3.PanelFromResource;
 import pcgen.gui3.behavior.EnabledOnlyWithCharacter;
 import pcgen.gui3.behavior.EnabledOnlyWithSources;
 import pcgen.gui3.dialog.ExportDialogController;
@@ -136,10 +136,9 @@ public final class PCGenToolBar
 
 	private void onExport(final ActionEvent actionEvent)
 	{
-		//GuiAssertions.assertIsNotJavaFXThread();
-		JFXPanelFromResource<ExportDialogController>
-				jfxPanelFromResource = new JFXPanelFromResource<>(ExportDialogController.class, "ExportDialog.fxml");
-		jfxPanelFromResource.showAsStage("Export a PC or Party");
+		PanelFromResource<ExportDialogController> panel =
+				new PanelFromResource<>(ExportDialogController.class, "ExportDialog.fxml");
+		panel.showAsStage("Export a PC or Party");
 	}
 
 	private void onPreferences(final ActionEvent actionEvent)
